[DPR2-347] Converted data table
