+++ src/dpr/components/data-table/view.njk
@@ -24,7 +24,7 @@
   {%- endfor %}
 
   {% if rows.length %}
-    <div id="table-container" class='table-container'>
+    <div id="table-container" class='table-container' data-dpr-module="data-table">
       {{ dprLoadingAnim() }}
       <div>
         <div class="govuk-grid-column">
